changes to approve the system
make 2 config files, one for test and one for real. add variable to
multiplayer - to know how the game ended and change the game table
according to that.
add pop-up to the player - in the first time he change the manue remind
him that the other player cant see whatever he change. change the places
that the data enter to the questionnaire table. add the text to the
translation in the game and logs.